fixing some line

+++ main.py
@@ -1,28 +1,27 @@
 # import functions and necessary library
+from functions import BalanceChecker
 from datetime import datetime
 from tqdm import tqdm
-import functions as fun
 import pandas as pd
-import time
+import time, os
 
 # import colorama libraries
 from colorama import init, Fore, Style
 init()
 
-def print_green(text):
-    print(Fore.GREEN + str(text) + Style.RESET_ALL)
-
 # timestamp the start
 start = time.time()
 
-# scraping the Crypto Explorer
+# assign the balance checker object
+bc = BalanceChecker("Explorers V3.csv")
+
+# check all the address balance
 tokens = []
-apis = fun.df2dict("Explorers V3.csv")
 today = datetime.now().strftime("%d/%m/%Y")
-for website in tqdm(apis,"Get Balance"):
-    api = apis[website]['api']
-    for address,symbols in apis[website]['source'].items():
-        new_tokens = fun.get_balance(website,api,address)
+for website in tqdm(bc.apis,"Get Balance"):
+    api = bc.apis[website]['api']
+    for address,symbols in bc.apis[website]['source'].items():
+        new_tokens = bc.get_balance(website,api,address)
         for symbol in symbols.split(","):
             try: balance = new_tokens[symbol]
             except: balance = 0
@@ -34,12 +33,16 @@ for website in tqdm(apis,"Get Balance"):
             }
             tokens.append(token)
 
-# convert data into dataframe
+# Concate to existing file
 outputfile = "Token Table.csv"
-old_data = pd.read_csv(outputfile)
-results = pd.DataFrame(tokens)
-results = pd.concat([old_data,results])
-results.to_csv(outputfile,index=False)
+if os.path.exists(outputfile):
+    old_data = pd.read_csv(outputfile)
+    results = pd.DataFrame(tokens)
+    results = pd.concat([old_data,results])
+    results.to_csv(outputfile,index=False)
+else:
+    results = pd.DataFrame(tokens)
+    results.to_csv(outputfile,index=False)
 
-# printout the elapse time
-print_green("Done in {:.2f} sec".format(time.time()-start))
+# printout the time elapse
+print(Fore.GREEN + "Done in {:.2f} sec".format(time.time()-start) + Style.RESET_ALL)
